NJU-ProjectN/navy-apps ics2023 initialized
NJU-ProjectN/navy-apps e8346911c80af39b63e47d8aba5878e43231ad63 libos,native: add comment for the format of /proc/dispinfo

+++ abstract-machine/klib/include/klib.h
@@ -31,7 +31,7 @@ void   free      (void *ptr);
 int    abs       (int x);
 int    atoi      (const char *nptr);
 void   itoa      (int value, char *string );
-void   hex2str   (unsigned int value, char *string, unsigned int upper);
+void   hex2str   (unsigned long int value, char *string, unsigned int upper);
 void   utostring (unsigned int value, char *string);
 
 // stdio.h
